refactor: does not delete folder on response, async process
