+++ src/task-manager.ts
@@ -54,7 +54,6 @@ export class TaskManager {
     serviceItemType: string,
     classifierApi: string,
     itemId: string,
-    fileStorage: string,
   ): CreateItemValidationTask {
     return new CreateItemValidationTask(
       member,
@@ -64,7 +63,6 @@ export class TaskManager {
       runner,
       serviceItemType,
       classifierApi,
-      fileStorage,
       { itemId },
     );
   }
