+++ src/task/create-item-validation-task.test.ts
@@ -29,7 +29,6 @@ const member = buildMember() as Member;
 const fTM = new FileTaskManager(DEFAULT_OPTIONS.serviceOptions, DEFAULT_OPTIONS.serviceMethod);
 const runner = new Runner();
 
-const fileStorage = './tmp';
 
 describe('Run detect bad words process', () => {
   const iVP = [SAMPLE_VALIDATION_PROCESS[0]];
@@ -63,7 +62,6 @@ describe('Run detect bad words process', () => {
       runner,
       ITEM_TYPE.LOCALFILE,
       MOCK_CLASSIFIER_API,
-      fileStorage,
       { itemId: input },
     );
     jest.spyOn(itemService, 'get').mockImplementation(async () => item as Item);
@@ -82,7 +80,6 @@ describe('Run detect bad words process', () => {
       runner,
       ITEM_TYPE.LOCALFILE,
       MOCK_CLASSIFIER_API,
-      fileStorage,
       { itemId: input },
     );
     jest.spyOn(itemService, 'get').mockImplementation(async () => item as Item);
